on click of search button input vals will be saved and data will be searched
All check boxes, text inputs values will be recored and search when the search button is clicked. Values will be reset to default when search is clicked. Default value for the search box is noodles, but this can be changed if the user wants. (The search being used needs a item to go into search)

Have not been able to test to make sure this works yet, will check later.